finhub-ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
langchain==0.3.12
langchain-ollama==0.2.1
langchain-core==0.3.25
langgraph==0.2.52
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
py-eureka-client==0.11.10
numpy==1.26.4
